Make Matchmaking Server a Configurable Setting
