+++ assets/default.game.yaml
@@ -55,6 +55,7 @@ main_menu:
   menu_width: 350
 
 default_settings:
+  matchmaking_server: localhost:8943
   player_controls:
     # Gamepad controls
     gamepad:
