+++ src/ui/main_menu/settings/sound.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+/// Render the sound settings UI
+pub fn sound_settings_ui(_ui: &mut egui::Ui, _game: &GameMeta) {
+    // This is un-reachable right now
+    todo!("Implement sound settings UI");
+}
